refactor: address PR review comments
- Memoize TokenTransferModalContext value to avoid unnecessary re-renders
- Move findCoinData and parseRawEventsForBalanceChanges to shared
  Transaction/utils module to reduce bundle coupling
- Add aria-label to modal close IconButton for accessibility
